add: bonus part

+++ utils/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: iammar <iammar@student.1337.ma>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 18:15:09 by iammar            #+#    #+#             */
-/*   Updated: 2024/12/30 13:38:52 by iammar           ###   ########.fr       */
+/*   Updated: 2025/01/05 20:39:49 by iammar           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-long long	ft_atoi(const char *str)
+double	ft_atoi(const char *str)
 {
-	long long	result;
-	int			sign;
+	double	result;
+	int		sign;
 
 	result = 0;
 	sign = 1;
